Crate bl602_pac

Peripheral access API for BL602 microcontrollers (generated using svd2rust v0.17.0 (3960899 2020-10-11))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Modules

aon
AON.

cci
cci.

cks
cks.

dma
dma.

ef_ctrl
ef_ctrl.

ef_data_0
ef_data_0.

ef_data_1
ef_data_1.

generic
Common register and bit access and modify traits

glb
glb.

gpip
gpip.

hbn
HBN.

i2c
i2c.

ir
ir.

l1c
l1c.

pds
pds.

pwm
pwm.

rf
rf.

sec_dbg
sec_dbg.

sec_eng
sec_eng.

sf_ctrl
sf_ctrl.

spi
spi.

timer
timer.

tzc_nsec
tzc_nsec.

tzc_sec
tzc_sec.

uart
uart.

Structs

AON
AON.

CCI
cci.

CKS
cks.

DMA
dma.

EF_CTRL
ef_ctrl.

EF_DATA_0
ef_data_0.

EF_DATA_1
ef_data_1.

GLB
glb.

GPIP
gpip.

HBN
HBN.

I2C
i2c.

IR
ir.

L1C
l1c.

PDS
pds.

PWM
pwm.

Peripherals
All the peripherals

RF
rf.

SEC_DBG
sec_dbg.

SEC_ENG
sec_eng.

SF_CTRL
sf_ctrl.

SPI
spi.

TIMER
timer.

TZC_NSEC
tzc_nsec.

TZC_SEC
tzc_sec.

UART
uart.